Setup progress bar on top
